Move files and adjust imports
