--- conda_classic_solver/solver.py
@@ -1,12 +1,0 @@
-# Copyright (C) 2022 Anaconda, Inc
-# Copyright (C) 2023 conda
-# SPDX-License-Identifier: BSD-3-Clause
-"""
-The classic solver implementation.
-"""
-
-from conda.core.solve import Solver
-
-
-class ClassicSolver(Solver):
-    pass
