+++ tests/test_logic.py
@@ -3,9 +3,9 @@
 from itertools import chain, combinations, permutations, product
 
 import pytest
-
-from conda.common.logic import FALSE, TRUE, Clauses, minimal_unsatisfiable_subset
 from conda.testing.helpers import raises
+
+from conda_classic_solver.logic import FALSE, TRUE, Clauses, minimal_unsatisfiable_subset
 
 # These routines implement logical tests with short-circuiting
 # and propagation of unknown values:
